Upgrade report template.
Add draft mode (pdf -> png).
Add fast mode.
Update README.
Remove unnecessary packages and files.
Add graphics.

+++ document/build.sh
@@ -9,17 +9,21 @@ CWD=$(pwd)
 INTERACTION=nonstopmode
 OUTDIR=dist
 JOBNAME=report
-PREVIEW=""
-OPEN=false
+DRAFT=""
+FAST=false
 SYNCTEX=1
 
-usage() { echo "Usage: $0 [-o]" 1>&2; exit 1; }
+usage() { echo "Usage: $0 [-f -d]" 1>&2; exit 1; }
 
-while getopts "o" o; do
+while getopts "df" o
+do
 	case "${o}" in
-		o)
-			PREVIEW="\def\preview{true}"
-			OPEN=true
+		f)
+			FAST=true
+			;;
+		d)
+			DRAFT="\def\draft{true}"
+			JOBNAME="$JOBNAME-draft"
 			;;
 		*)
 			usage
@@ -28,27 +32,32 @@ while getopts "o" o; do
 done
 shift $((OPTIND-1))
 
-rm -rf ${OUTDIR}/*
-mkdir -p ${OUTDIR}
+if [ $FAST == false ]
+then
+	mkdir -p ${OUTDIR}/tmp
+	cp ${OUTDIR}/*.pdf ${OUTDIR}/tmp || true
+	rm -f ${OUTDIR}/*.*
+	cp ${OUTDIR}/tmp/*.pdf ${OUTDIR} || true
+	rm -rf ${OUTDIR}/tmp
+	mkdir -p ${OUTDIR}/figures
+fi
 
 if [ -n "$CI_BUILD_REF" ];
 then
 	printf "\providecommand{\\\version}{%s}" $(echo $CI_BUILD_REF | cut -c1-8) > version.tex
 fi
 
-latexmk -cd -dvi- -f -pdf -ps- -time \
-	-jobname=$JOBNAME \
-	-outdir=$OUTDIR \
+latexmk -cd -dvi- -f -pdf -ps- -time -shell-escape \
+	-jobname=${JOBNAME} \
+	-outdir=${OUTDIR} \
 	--synctex=${SYNCTEX} \
 	--interaction=${INTERACTION} \
-	-pdflatex='pdflatex %O "\def\dummy{} '${PREVIEW}' \input %S "' \
+	-pdflatex='pdflatex %O "\def\dummy{} '${REVIEW}' '${DRAFT}' \input %S "' \
 	main
 
-rm -f ${OUTDIR}/*.{aux,log,out,xwm,toc,lof,lot,bib,bbl,bcf,blg,xml,fls,fdb_latexmk}
+if [ $FAST == false ]
+then
+	rm -f ${OUTDIR}/*.{aux,log,out,xwm,toc,lof,lot,bib,bbl,bcf,blg,xml,fls,fdb_latexmk}
+fi
 
 echo "Done."
-
-if [ "$OPEN" == true ]
-then
-	open ${OUTDIR}/${JOBNAME}.pdf
-fi
